requirements.txt
Flask==2.2.2
Flask-PyMongo==2.3.0
Flask-OAuthlib==0.9.6
python-dotenv==0.19.2
requests==2.28.1
pymongo==4.1.1
Werkzeug==2.2.2
gunicorn==20.1.0
dnspython